Refactor Normalize.tsx, update card_data.json
