+++ .gitignore
@@ -49,3 +49,6 @@ __pycache__/
 
 # dataset downloads
 /src/ai_dev/datasets/pokemon/data/
+
+# workspace settings
+.vscode/
